Merge Group4mastercopy2 in master

+++ Controllers/CategoriesController.cs
@@ -1,7 +1,9 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Drawing.Drawing2D;
 using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
@@ -10,6 +12,7 @@ using Supermarket.Models;
 
 namespace Supermarket.Controllers
 {
+    [Authorize(Roles = "Stock Administrator, Stock Operator")]
     public class CategoriesController : Controller
     {
         private readonly SupermarketDbContext _context;
@@ -20,11 +23,83 @@ namespace Supermarket.Controllers
         }
 
         // GET: Categories
-        public async Task<IActionResult> Index()
+        public async Task<IActionResult> Index(int page = 1, string name = "")
         {
-              return _context.Category != null ? 
-                          View(await _context.Category.ToListAsync()) :
-                          Problem("Entity set 'SupermarketDbContext.Category'  is null.");
+            var categories = _context.Category.AsQueryable();
+
+            if (name != "")
+            {
+                categories = categories.Where(x => x.Name.Contains(name));
+            }
+
+            PagingInfoProduct paging = new PagingInfoProduct
+            {
+                CurrentPage = page,
+                TotalItems = await categories.CountAsync(),
+            };
+
+            if (paging.CurrentPage <= 1)
+            {
+                paging.CurrentPage = 1;
+            }
+            else if (paging.CurrentPage > paging.TotalPages)
+            {
+                paging.CurrentPage = paging.TotalPages;
+            }
+
+            var vm = new CategoriesViewModel
+            {
+                Categories = await categories
+                    .OrderBy(b => b.Name)
+                    .Skip((paging.CurrentPage - 1) * paging.PageSize)
+                    .Take(paging.PageSize)
+                    .ToListAsync(),
+                PagingInfoProduct = paging,
+                SearchName = name,
+            };
+
+
+            return View(vm);
+            //var categories = from b in _context.Category select b;
+
+            //if (name != "")
+            //{
+            //    categories = categories.Where(x => x.Name.Contains(name));
+            //}
+
+            //PagingInfoProduct paging = new PagingInfoProduct
+            //{
+            //    CurrentPage = page,
+            //    TotalItems = await categories.CountAsync(),
+            //};
+
+            //if (paging.CurrentPage <= 1)
+            //{
+            //    paging.CurrentPage = 1;
+            //}
+            //else if (paging.CurrentPage > paging.TotalPages)
+            //{
+            //    paging.CurrentPage = paging.TotalPages;
+            //}
+
+            //var vm = new CategoriesViewModel
+            //{
+            //    Categories = await categories
+            //        .OrderBy(b => b.Name)
+            //        .Skip((paging.CurrentPage - 1) * paging.PageSize)
+            //        .Take(paging.PageSize)
+            //        .ToListAsync(),
+            //    PagingInfoProduct = paging,
+            //    SearchName = name,
+            //};
+
+            ////var totalCategories = await categories.CountAsync()
+            //ViewBag.totalCategories = vm.PagingInfoProduct.TotalItems;
+
+            //return View(vm);
+            //return _context.Category != null ? 
+            //              View(await _context.Category.OrderBy(c => c.Name).ToListAsync()) :
+            //              Problem("Entity set 'SupermarketDbContext.Category'  is null.");
         }
 
         // GET: Categories/Details/5
@@ -46,6 +121,7 @@ namespace Supermarket.Controllers
         }
 
         // GET: Categories/Create
+        [Authorize(Roles = "Stock Administrator")]
         public IActionResult Create()
         {
             return View();
@@ -56,6 +132,7 @@ namespace Supermarket.Controllers
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
+        [Authorize(Roles = "Stock Administrator")]
         public async Task<IActionResult> Create([Bind("CategoryId,Name")] Category category)
         {
             if (ModelState.IsValid)
@@ -68,6 +145,7 @@ namespace Supermarket.Controllers
         }
 
         // GET: Categories/Edit/5
+        [Authorize(Roles = "Stock Administrator")]
         public async Task<IActionResult> Edit(int? id)
         {
             if (id == null || _context.Category == null)
@@ -88,6 +166,7 @@ namespace Supermarket.Controllers
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
+        [Authorize(Roles = "Stock Administrator")]
         public async Task<IActionResult> Edit(int id, [Bind("CategoryId,Name")] Category category)
         {
             if (id != category.CategoryId)
@@ -119,6 +198,7 @@ namespace Supermarket.Controllers
         }
 
         // GET: Categories/Delete/5
+        [Authorize(Roles = "Stock Administrator")]
         public async Task<IActionResult> Delete(int? id)
         {
             if (id == null || _context.Category == null)
@@ -139,6 +219,7 @@ namespace Supermarket.Controllers
         // POST: Categories/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
+        [Authorize(Roles = "Stock Administrator")]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             if (_context.Category == null)
@@ -146,7 +227,13 @@ namespace Supermarket.Controllers
                 return Problem("Entity set 'SupermarketDbContext.Category'  is null.");
             }
             var category = await _context.Category.FindAsync(id);
-            if (category != null)
+            var product = await _context.Product.Where(p => p.BrandId == id).ToListAsync();
+            if (product != null)
+            {
+                ModelState.AddModelError(string.Empty, "There is a product with this category, can't be deleted.");
+                return View(category);
+            }
+            else if (category != null)
             {
                 _context.Category.Remove(category);
             }
